refactor - put typedefs into types.hpp and more utility functions into utils
