+++ mogp_gpu/src/gp_gpu.hpp
@@ -21,162 +21,11 @@ These in turn use CUDA kernels defined in the file cov_gpu.cu
 #include <sstream>
 #include <assert.h>
 #include <stdexcept>
-#include <cuda_runtime.h>
-#include <cublas_v2.h>
-#include <cusolverDn.h>
-#include <thrust/device_ptr.h>
-#include <thrust/device_malloc.h>
-#include <thrust/device_free.h>
-#include <thrust/device_vector.h>
-#include <thrust/functional.h>
-#include <thrust/transform_reduce.h>
-#include <thrust/copy.h>
-#include <cub/cub.cuh>
 
-#include <Eigen/Dense>
-
-#include "strided_range.hpp"
-#include "kernel.hpp"
-//#include "cov_gpu.hpp"
 #include "util.hpp"
-
-#define WARP_SIZE 32
-#define FULL_MASK 0xffffffff
-
-#define USE_SHUFFLE_SUM_IMPL 0
-
-// ----------------------------------------------
-// ----   Utility functions ---------------------
-
-/// Extract the raw pointer from a device vector
-template <typename T>
-T *dev_ptr(thrust::device_vector<T>& dv)
-{
-    return dv.data().get();
-}
+#include "kernel.hpp"
 
 
-/// Fail if a recent cusolver call did not succeed
-inline void check_cusolver_status(cusolverStatus_t status, int info_h)
-{
-    if (status || info_h) {
-	std::string msg;
-	std::stringstream smsg(msg);
-	smsg << "Error in potrf: return code " << status << ", info " << info_h;
-	throw std::runtime_error(smsg.str());
-    }
-}
-
-/// Can a usable CUDA capable device be found?
-bool have_compatible_device(void);
-
-// ----------------------------------------------
-
-
-// --------------------------------------------------------
-// CUDA kernel for summing log diagonal elements of a matrix.
-// Two methods implemented - warp reduction or cub::DeviceReduce
-
-#if USE_SHUFFLE_SUM_IMPL
-// Implementation of sum_log_diag using warp reduction,
-// based on https://devblogs.nvidia.com/faster-parallel-reductions-kepler/
-//
-// Will fail for N > 1024
-
-
-__inline__ __device__
-double warp_reduce_sum(double val)
-{
-    for (int offset = WARP_SIZE/2; offset > 0; offset /= 2)
-        val += __shfl_down_sync(FULL_MASK, val, offset);
-    return val;
-}
-
-__global__
-void sum_log_diag_kernel(int N, double *A, double *result)
-{
-    int i = threadIdx.x;
-
-    static __shared__ double work[WARP_SIZE];
-    double log_diag = 0.0;
-
-    if (i < WARP_SIZE) work[i] = 0.0;
-    if (i < N) {
-        log_diag = log(A[i * (N+1)]);
-    }
-
-    int laneIdx = i % WARP_SIZE;
-    int warpIdx = i / WARP_SIZE;
-
-    __syncthreads();
-    log_diag = warp_reduce_sum(log_diag);
-
-    if (laneIdx == 0) work[warpIdx] = log_diag;
-
-    __syncthreads();
-
-    log_diag = work[laneIdx];
-
-    if (warpIdx == 0) log_diag = warp_reduce_sum(log_diag);
-
-    if (i == 0) *result = 2.0 * log_diag;
-}
-
-//
-// The unused arguments are needed for the CUB implementation
-void sum_log_diag(int N, double *A, double *result, double *, size_t)
-{
-    // The number of thread blocks *must* be 1
-    // N can be at most 1024 (WARP_SIZE**2)
-    int thread_blocks = 1;
-    int threads_per_block = N;
-    sum_log_diag_kernel<<<thread_blocks, threads_per_block>>>(N, A, result);
-}
-
-#else
-// ----------------------------------------
-
-// Implementation of sum_log_diag using cub::DeviceReduce
-
-struct LogSq : public thrust::unary_function<double, double>
-{
-    __host__ __device__ double operator()(double x) const { return 2.0 * log(x); }
-};
-
-void sum_log_diag(int N, double *A, double *result, double *work, size_t work_size)
-{
-    auto transform_it = thrust::make_transform_iterator(A, LogSq());
-    auto sr = make_strided_range(transform_it, transform_it + N*N, N+1);
-
-    cub::DeviceReduce::Sum(work, work_size, sr.begin(), result, N);
-}
-
-// Implementation of trace using cub::DeviceReduce
-
-void trace(int N, double *A, double *result, double *work, size_t work_size)
-{
-    auto sr = make_strided_range(A, A + N*N, N+1);
-    cub::DeviceReduce::Sum(work, work_size, sr.begin(), result, N);
-}
-
-// ----------------------------------------
-#endif // USE_SHUFFLE_SUM_IMPL
-
-
-// ----------------------------------------
-// ------- Some useful typedefs for vectors and matrices
-
-typedef int obs_kind;
-typedef typename Eigen::Matrix<REAL, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> mat;
-typedef typename Eigen::Ref<Eigen::Matrix<REAL, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> > mat_ref;
-typedef typename Eigen::Matrix<REAL, Eigen::Dynamic, 1> vec;
-typedef typename Eigen::Ref<Eigen::Matrix<REAL, Eigen::Dynamic, 1> > vec_ref;
-// ----------------------------------------
-// enum to allow the python code to select the type of "nugget"
-enum nugget_type {NUG_ADAPTIVE, NUG_FIT, NUG_FIXED};
-
-// enum to allow python code to select Kernel function
-enum kernel_type {SQUARED_EXPONENTIAL, MATERN52};
 
 
 //////////////////////////////////////////////////////////////////////////////////////
